Remove `cmd` from `useAddApp` hook and move to component `Automatically`
