+++ packages/teleport/src/Apps/AddApp/Automatically/Automatically.test.tsx
@@ -1,21 +1,21 @@
 import React from 'react';
 import { fireEvent, render, screen } from 'design/utils/testing';
-import Automatically from './Automatically';
+import Automatically, { createAppBashCommand } from './Automatically';
 
 test('render command only after form submit', async () => {
-  const cmd = 'sudo some-command';
+  const token = 'token';
   render(
     <Automatically
-      cmd={cmd}
+      token={token}
       attempt={{ status: 'success' }}
       onClose={() => {}}
-      onSubmit={() => {}}
       onCreate={() => Promise.resolve(true)}
       expires=""
     />
   );
 
   // initially, should not show the command
+  let cmd = createAppBashCommand(token, '', '');
   expect(screen.queryByText(cmd)).toBeNull();
 
   // set app name
@@ -32,5 +32,17 @@ test('render command only after form submit', async () => {
   screen.getByRole('button', { name: /Generate Script/i }).click();
 
   // after form submission should show the command
+  cmd = createAppBashCommand(token, 'app-name', 'https://gravitational.com');
   expect(screen.queryByText(cmd)).not.toBeNull();
+});
+
+test('app bash command encoding', () => {
+  const token = '86';
+  const appName = 'jenkins';
+  const appUri = `http://myapp/test?b='d'&a="1"&c=|`;
+
+  const cmd = createAppBashCommand(token, appName, appUri);
+  expect(cmd).toBe(
+    `sudo bash -c "$(curl -fsSL 'http://localhost/scripts/86/install-app.sh?name=jenkins&uri=http%3A%2F%2Fmyapp%2Ftest%3Fb%3D%27d%27%26a%3D%221%22%26c%3D%7C')"`
+  );
 });
